+++ import_sorter/args.py
@@ -1,26 +1,45 @@
 import sys
-import glob
+import shlex
 import tomllib
 from pathlib import Path
 from dataclasses import dataclass
+from pathspec import GitIgnoreSpec
 from argparse import ArgumentParser
-from typing import Self, TextIO, Literal, Sequence
+from typing import Self, Iterator, Sequence
 
 
 @dataclass
+class Source:
+    content: str
+
+
+@dataclass(kw_only=True)
 class Args:
-    file: str
+    files: list[str]
+    exclude: list[str]
     groups: list[str]
-    format: str | None
+    format: list[str]
     config: str | None
-    recursive: bool = False
+    no_config: bool
+    as_json: bool
+    quiet: bool
 
     def __post_init__(self):
         if self.config:
             self._load_toml(self.config, [])
 
-        self._load_toml("import-sorter.toml", [])
-        self._load_toml("pyproject.toml", ["tool", "import-sorter"])
+        if not self.no_config:
+            self._load_toml("import-sorter.toml", [])
+            self._load_toml("pyproject.toml", ["tool", "import-sorter"])
+
+        self.format = shlex.split(self.format[0]) if len(self.format) == 1 else self.format
+
+    def _config_list(self, value: list[str] | str | None) -> list[str]:
+        if isinstance(value, str):
+            return shlex.split(value)
+        elif isinstance(value, list):
+            return value
+        return []
 
     def _load_toml(self, file: str, path: list[str]):
         try:
@@ -34,11 +53,9 @@ class Args:
         for key in path:
             config = config.get(key, {})
 
-        if not self.groups:
-            self.groups = config.get("groups", [])
-
-        if not self.format:
-            self.format = config.get("format")
+        self.exclude.extend(self._config_list(config.get("exclude")))
+        self.groups = self.groups or self._config_list(config.get("groups"))
+        self.format = self.format or self._config_list(config.get("format"))
 
     @classmethod
     def parse(cls, args: Sequence[str] | None = None) -> Self:
@@ -50,37 +67,120 @@ class Args:
             prog = exec_file.name
 
         parser = ArgumentParser(prog)
-
-        parser.add_argument("file")
-        parser.add_argument("-g", "--groups", action="extend", nargs="+", default=[])
-        parser.add_argument("-f", "--format", default=None)
-        parser.add_argument("-c", "--config", default=None)
         parser.add_argument(
-            "-r", "--recursive", action="store_true", help="Format all Python files in directory recursively"
+            "files",
+            action="extend",
+            nargs="+",
+            default=[],
+            help=(
+                "List of files, folders or globs to format. "
+                "Specifying folders formats every file recursively, "
+                "write '<folder>/*.py' if you're not looking for that. "
+                "If '-' is specified, then the program reads from stdin until closed, "
+                "and writes to stdout the formatted contents. "
+                "Files with syntax errors are not formatted. "
+                "Cannot be specified in config."
+            ),
+        )
+        parser.add_argument(
+            "-x",
+            "--exclude",
+            action="extend",
+            nargs="+",
+            default=[],
+            help=(
+                ".gitignore like list of exclusions for multiple files. "
+                "Negative exclusions (!) only work for specified files. "
+                "eg.: '.venv' ignores every file inside .venv folders. "
+                "Can be specified in config, as a string or list of strings."
+            ),
+        )
+        parser.add_argument(
+            "-g",
+            "--groups",
+            action="extend",
+            nargs="+",
+            default=[],
+            required=False,
+            help=(
+                "Import group order for specific packages. eg.: '-g tomllib argparse' "
+                "group order: __future__ -> other imports -> tomllib imports -> argparse imports. "
+                "Can be specified in config, as a string or list of strings."
+            ),
+        )
+        parser.add_argument(
+            "-f",
+            "--format",
+            action="extend",
+            nargs="+",
+            default=[],
+            required=False,
+            help=(
+                "Format command arguments. "
+                "If the command starts with 'python', use the current executable. "
+                "The format should accept a stdio stream. eg.: '-f ruff format -'. "
+                "Can be specified in config, as a string or list of strings."
+            ),
+        )
+        parser.add_argument(
+            "-c",
+            "--config",
+            default=None,
+            help=(
+                "Config toml file to load arguments from. eg.: '-c import-sorter-args.toml' "
+                "Cannot be specified in config."
+            ),
+        )
+        parser.add_argument(
+            "-n",
+            "--no-config",
+            action="store_true",
+            default=False,
+            help=(
+                "Disables the default behaviour of loading from specific configs. "
+                "By default, the program always tries to read from the current directory's specified config file at "
+                "root, then from `pyproject.toml` at [tool.import-sorter], then from `import-sorter.toml` at root. "
+                "If this argument is passed, only the specified config file will be used (if specified). "
+                "Cannot be specified in config."
+            ),
+        )
+        print_group = parser.add_mutually_exclusive_group()
+        print_group.add_argument(
+            "-j",
+            "--as-json",
+            action="store_true",
+            default=False,
+            help="Print status as json strings separated by newlines. Cannot be specified from config.",
+        )
+        print_group.add_argument(
+            "-q",
+            "--quiet",
+            action="store_true",
+            default=False,
+            help="Don't print any status. Cannot be specified from config",
         )
 
         return cls(**parser.parse_args(args).__dict__)
 
-    def open_file(self, mode: Literal["r", "w"]) -> TextIO:
-        if self.file != "-":
-            return open(self.file, mode, encoding="utf-8")
+    def list_files(self) -> Iterator[str]:
+        exclude_spec = GitIgnoreSpec.from_lines(self.exclude)
 
-        return sys.stdin if mode == "r" else sys.stdout
+        for f in self._iterate_files(self.files):
+            if not exclude_spec.match_file(f):
+                yield f
 
-    def get_python_files(self) -> list[str]:
-        if self.file == "-":
-            return ["-"]
+    def _iterate_files(self, files: list[str]):
+        for file in files:
+            if file == "-":
+                yield file
+                continue
 
-        file_path = Path(self.file)
+            path = Path(file)
 
-        if file_path.is_file():
-            return [str(file_path)]
-        elif file_path.is_dir():
-            if self.recursive:
-                return [str(p) for p in file_path.rglob("*.py")]
-            else:
-                return [str(p) for p in file_path.glob("*.py")]
-        else:
-            matches = glob.glob(self.file, recursive=self.recursive)
-            python_files = [f for f in matches if f.endswith(".py")]
-            return python_files or [self.file]
+            if path.is_file() and path.suffix == ".py":
+                yield str(path.resolve())
+                continue
+
+            for subpath in path.glob("**/*.py") if path.is_dir() else Path().glob(file):
+                if subpath.suffix == ".py":
+                    yield str(subpath.resolve())