+++ import_sorter/sorting.py
@@ -27,26 +27,21 @@ class ImportSorter:
 
         new_source = ""
 
-        try:
-            *first_groups, last_group = [group for group in self.groups.values() if group]
-        except ValueError:
-            # No imports
-            return self.source
+        for group in self.groups.values():
+            if not group:
+                continue
 
-        for group in first_groups:
             for import_stmt in self._sort_imports(group.values()):
-                new_source += ast.unparse(import_stmt)
                 new_source += "\n"
+                new_source += ast.unparse(import_stmt)
             new_source += "\n"
 
-        for import_stmt in self._sort_imports(last_group.values()):
-            new_source += ast.unparse(import_stmt)
-            new_source += "\n"
+        new_source = new_source.strip() + "\n"
 
         for line in self.source.splitlines(keepends=True)[last_line:]:
             new_source += line
 
-        return new_source
+        return new_source.lstrip()
 
     def _read_stmt(self, stmt: ast.stmt) -> bool | None:
         if isinstance(stmt, ast.Import):
@@ -103,3 +98,7 @@ class ImportSorter:
         return len(unparsed), len(module), unparsed, module
 
     # endregion
+
+
+def sort_imports(source: str, groups: Sequence[str]):
+    return ImportSorter(source, groups).sort()