+++ vscode-extension/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the "Python Import Sorter" extension will be documented i
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-06-09
+
+### Added
+- A way to format every python file in a workspace
+- The `python-import-sorter.exclude` setting
+- The need to have the [Python](https://marketplace.visualstudio.com/items?itemName=ms-python.python) extension
+
+### Removed
+- The need to have [Python Environments](https://marketplace.visualstudio.com/items?itemName=ms-python.vscode-python-envs) extension
+
 ## [1.0.0] - 2025-04-07
 
 ### Added